Fix broken provided DevicePreviewStyle
Currently when you make your own DevicePreviewStyle and pass it in to the constructor of DevicePreivew via its style property. It does not actually do anything. It always uses the default one via the inherited style. If you instead only use the inherited style if passed in widget.style is null, this issue is resolved and DevicePreview uses the provided DevicePreviewStyle one instead, if one is provided.